refactor: extract onKeydown, handleKeydown, handleUpdate to separate file
